opt: experimental/benchmarks/web_tooling_benchmark      exchange the results of benchmark web_tooling_benchmark from float to int
opt: experimental/driver/benchmarks.py
     add a new opton for octane & web_tooling_benchmark only:
     --runs: run times of benchmark octane or web_tooling_benchmark

+++ experimental/benchmarks/web_tooling_benchmark/postdata.py
@@ -30,7 +30,7 @@ if __name__ == '__main__':
         line = f.readline()
         while line:
              if "Geometric mean" in line:
-                  val = float(line.split()[-2])*100
+                  val = int(float(line.split()[-2])*100)
                   print val
                   arr.append(val)
              line = f.readline()
